checks if directory exists

+++ upload_script.py
@@ -13,16 +13,24 @@ s3resource.Bucket('test-charissa').put_object(Key='word4-word5-word6', Body=data
 path = "../boop-bawp-beep"
 bucketname = "test-charissa"
 
+
+
 def uploadDirectory(path, bucketname):
+    new_upload = False
     for root, dirs, files in os.walk(path):
         for dir in dirs:
             directory_name = os.path.join(root, dir) + "/"
             directory_name = directory_name.replace("../", "")
-            s3client.put_object(
-                Bucket='test-charissa',
-                Body='',
-                Key=directory_name
-            )
+            root_folder_name = directory_name.split('/')
+            root_folder_name = root_folder_name[0]
+            if root_folder_name not in existing_folders:
+                new_upload = True
+                s3client.put_object(
+                    Bucket='test-charissa',
+                    Body='',
+                    Key=directory_name
+                )
+        # if new_upload:
         for file in files:
             filename = os.path.join(file)
             file_with_path = os.path.join(root, file)
@@ -34,10 +42,22 @@ def uploadDirectory(path, bucketname):
                 Body=data,
                 Key=new_directory
             )
-
+            new_upload = False
 
 # find names of directories in current path
-s3client.list_objects(Bucket='test-charissa')
+folders = s3client.list_objects(Bucket='test-charissa')
+
+existing_folders = set([])
+
+for thing in s3resource.Bucket('test-charissa').objects.all():
+    location = thing.key.split('/')
+    location = location[0]
+    existing_folders.add(location)
+
+print existing_folders
+
+
+
 # find names of directories on s3
 # if name in current path does not exist on s3, then upload that directory
 uploadDirectory(path, bucketname)
